==> Part 1 - Basic Cleaning.ipynb <==
import pandas as pd
import numpy as np
import matplotlib as plt
import seaborn as sns
%matplotlib inline

df = pd.read_csv('Data/pitchesv1.csv')

df.shape

# Dropping features with all nulls
df = df.dropna(axis=1, thresh=2)
df.shape

# Dropping irrelevant features
dropped_columns = ['pitch_tfs_zulu', 'pitch_tfs', 'final_outs', 'at_bat_des', 
                   'type_confidence', 'modified_by', 'modified_at', 'added_at',
                  'date', 'sv_id', 'created_at', 'start_tfs', 'start_tfs_zulu', 'year']
df = df.drop(dropped_columns, axis = 1)
df.shape

df.isnull().sum().sort_values(ascending=False)[1:30]

# Turning Features that are objects into binary features IE 'R' or 'L' to 1 and 0
df['score'] = df.score.fillna(0)
df['on_1b'] = df.on_1b.fillna(0)
df['on_2b'] = df.on_2b.fillna(0)
df['on_3b'] = df.on_3b.fillna(0)
df['score'] = (df['score'] != 0).astype(int)
df['on_1b'] = (df['on_1b'] > 0).astype(int)
df['on_2b'] = (df['on_2b'] > 0).astype(int)
df['on_3b'] = (df['on_3b'] > 0).astype(int)
df['score'] = (df['score'] != 0).astype(int)
df['right_side_plate'] = (df['stand'] == 'R').astype(int)
df['right_handed_pitcher'] = (df['p_throws'] == 'R').astype(int)
df = df.drop(['stand','p_throws'], axis = 1)
df.shape

#Removing null rows for pitch_type (outcome variable)
df = df[pd.notnull(df['pitch_type'])]
df.isnull().sum().sort_values(ascending=False)[0:10]

# Turning Batter Height from Object into Integer
def parse_ht(ht):
    ht_ = ht.split("-")
    ft_ = int(ht_[0])
    in_ = int(ht_[1].replace("\"",""))
    return (12*ft_) + in_

df['batter_height_inches'] = (df['b_height'].apply(lambda x:parse_ht(x))).astype(int)
df = df.drop('b_height', axis = 1)

df.to_csv('Data/pitchesv2.csv', index=False)
