==> Part 2 - Creating Pitch Type Percentage.ipynb <==
import pandas as pd
import numpy as np
import matplotlib as plt
import seaborn as sns
%matplotlib inline

df = pd.read_csv('Data/pitchesv2.csv')

df.pitch_type.value_counts()

# Dropping pitch AB and UN due to low value counts
# Replacing FO and FF due to 
# https://www.fangraphs.com/library/pitch-type-abbreviations-classifications/

df = df[df['pitch_type'] != 'AB']
df = df[df['pitch_type'] != 'UN']
df = df.replace('FF', 'FA')
df = df.replace('FO', 'PO')

# Checking to see if above cell worked
df.pitch_type.value_counts()

pitchers_id_type = df[['pitcher_id', 'pitch_type']]
pitchers_id_type.head()

# Grouping by pitcher id and then seeing the percentage a certain ball is thrown
# in 2011 season. Should be really helpful by allowing the model to not even consider
# certain choices like 'AB' type pitch for the first five pitcher ids

pitchers_id_type = pitchers_id_type.rename(index=str, 
                                           columns={'pitch_type' : 'percentage'})
pitchers_id_type = pd.get_dummies(pitchers_id_type)
pitchers_id_type = pitchers_id_type.groupby(['pitcher_id']).mean()
pitchers_id_type.head()

df.to_csv('Data/pitchesv2.csv', index=False)
pitchers_id_type.to_csv('Data/percentages.csv')
