==> Part 3 - Creating "Previous Features".ipynb <==
import pandas as pd
import numpy as np
import matplotlib as plt
import seaborn as sns
%matplotlib inline

df = pd.read_csv('Data/pitchesv2.csv')
df.head()

# Dropping these variables to make it easier on my computer but 
# I would do previous events for all of these variables
# Kept the most important variables based on my baseball intuition
previous_pitch_columns = ['type', 'x', 'y', 'pitch_type', 'start_speed', 'event']
df = df.drop(['final_balls', 'final_strikes', 'event4', 'is_final_pitch', 'pitch_des', 
              'end_speed', 'pfx_x', 'pfx_z', 'px', 'pz', 'x0', 'z0', 'y0', 'vz0', 
              'vy0', 'az', 'ay', 'break_length', 'break_angle', 'break_y', 
              'spin_dir', 'spin_rate', 'nasty', 'vx0', 'ax', 'sz_top', 'sz_bot',
             'event2', 'event3', 'cc'], axis=1)

#Creating Previous Pitch Features
pitching_groupby = ['pitcher_id', 'at_bat_num', 'batter_id', 'game_pk', 'inning']
df['previous_pitch'] = df.groupby(pitching_groupby)['pitch_type'].shift(1)
df['previous_pitch'] = df['previous_pitch'].fillna('First_Pitch')

df['previous_type'] = df.groupby(pitching_groupby)['type'].shift(1)
df['previous_type'] = df['previous_type'].fillna('First_Pitch')
df = df.drop('type', axis = 1)


# Changing More Previous Pitch features but not using the prefix previous
previous_pitch_columns = ['x', 'y', 'start_speed', 'zone']

for column in previous_pitch_columns:
    df[column] = df.groupby(pitching_groupby)[column].shift(1)
    df[column] = df[column].fillna(df[column].mean())

# Creating Previous At Bat Features
at_bat_groupby = ['pitcher_id', 'inning','game_pk']
df['previous_at_bat_event'] = df.groupby(at_bat_groupby)['event'].shift(1)
df['previous_at_bat_event'] = df['previous_at_bat_event'].fillna('New_Inning')
df['previous_score'] = df.groupby(at_bat_groupby)['score'].shift(1)
df['previous_score'] = df['previous_score'].fillna('New_Inning')

# Dropping event and score because I've created previous event and previous score
df = df.drop(['event', 'score'], axis = 1)
df.to_csv('Data/pitchesv3.csv', index = False)
