==> Part 4 - Merging Dataframes.ipynb <==
import pandas as pd
import numpy as np
import matplotlib as plt
import seaborn as sns

df = pd.read_csv('Data/pitchesv3.csv')
pitching_percentages = pd.read_csv('Data/percentages.csv')

df.shape

pitching_percentages.shape

# Merging dataframes and checking if columns added up
df = pd.merge(pitching_percentages, df, on= ['pitcher_id'])
df.shape

# Checking if percentage features are different based on pitcher_id
df.iloc[200:205,:5]

df.iloc[12000:12005,:5]

# Dropping irrelevant features
df = df.drop(['uid', 'pitch_id'], axis=1)

# Dropped because it added too many features and did not seem like it would add much, similar to previous_event
df = df.drop('previous_at_bat_event', axis = 1)

# Getting dummies for everything but pitch_type
df = pd.get_dummies(df, columns=['previous_pitch', 'previous_type', 'previous_score'])

# Dropping previous score 0.0 because it's inverse of previous score 1.0
df = df.drop('previous_score_0.0', axis=1)
df.to_csv('modelingv2.csv', index=False)
