==> Part 10 - Grid Search and Applying to the Test Set.ipynb <==
# Picked only a couple of params, would do more with more time
rfc_params = {'n_estimators' : [5, 10], 'max_depth' : [5, None], 'random_state' : [12]}
ridge_params = {'alpha' : [1, 2], 'max_iter' : [5, None], 'random_state' :[12]}

df_train, df_test = train_test_split(df, test_size=0.4, random_state = 12)
X_train = df_train.loc[:, ~(df_train.columns).isin(['pitch_type'])]
Y_train = df_train['pitch_type']
X_test = df_test.loc[:, ~(df_test.columns).isin(['pitch_type'])]
Y_test = df_test['pitch_type']

# Grid Search RFC
rfc = RandomForestClassifier()
GSrfc = GridSearchCV(rfc, rfc_params)
GSrfc.fit(X_train, Y_train)
GSrfc.best_params_

# Grid Search Ridge
ridge = RidgeClassifier()
GSridge = GridSearchCV(ridge, ridge_params)
GSridge.fit(X_train, Y_train)
GSridge.best_params_

bestrfc = RandomForestClassifier(max_depth= None, n_estimators= 10, random_state= 12)

bestridge = RidgeClassifier(alpha=1, max_iter = 5, random_state=12)

The Model I am choosing is Ridge Classifier due to the higher scores across the board.

def test_metrics(model, x_train, y_train, x_test, y_test):
    model.fit(x_train, y_train)
    y_pred = model.predict(x_test)
    accuracy = accuracy_score(y_test, y_pred)
    macro = f1_score(y_test, y_pred, average='macro')
    micro = f1_score(y_test, y_pred, average='micro')
    weighted = f1_score(y_test, y_pred, average='weighted')
    print(model)
    print('Accuracy:', accuracy)
    print('F1 Macro:', macro)
    print('F1 Micro:', micro)
    print('F1 Weighted:', weighted)
    print(classification_report(y_test, y_pred))

test_metrics(bestridge, X_train, Y_train, X_test, Y_test)

The final model performed roughly the same as the test set. The model is really good at predicting Intentional Balls and Knuckleballs. It also can't predict an Ephesus pitch very well. However even casual fans can figure out when intentional balls are about to be thrown so this finding isn't very insightful. Some ideas I have are to drop features like team_id and maybe aggregate some of the outcome variables like the Two-Seam Fastball with the Four-Seam Fastball. I see a correlation with pitching choice and the pitches thrown in the game.  In the Random Forest model pitching and at bats were both very important features and both are related to length time played and pitches thrown.
